Update ContentDb to set template defaults during build process
This was causing a bug where defaults were not being set for components that
weren't part of the root template.

For example, with this template heirarchy:

  - base_template : { component_one }
    - leaf_template : { component_two }
      - entity_template

the default values for `component_two` were not being set at all. This causes
some weirdness (and potential crashes) when adding components to leaf templates.