Fix `MoveSpriteBehaviour` so it works with multiple effects
The previous version worked /okay/ on its own, but when two movements were
applied to the same sprite it would move all over the place.

This version calculates the delta to move the sprite and applies it each frame,
and no longer uses the sprite's x/y co-ordinates for calculating offsets.