Add `pangolin.actions` module
This is adds the `BackgroundActionsService`, which is used to perform actions in
the background that don't really need threads.

For example, it can be used to fade the screen in/out or change the volume of
music over a set amount of time.

Graphics-related actions are handled by `pangolin.gfx_actions`.